Got the test in question to pass. This change adequately protects us from docker.
